change models, views and tests into package
