+++ kuhub/admin.py
@@ -1,6 +1,6 @@
 from django.contrib import admin
 
-from .models import Blog, Tag, Comment, BlogReport, CommentReport
+from kuhub.models import Blog, Tag, Comment, BlogReport, CommentReport
 
 admin.site.register(Tag)
 admin.site.register(Blog)
