+++ kuhub/models/__init__.py
@@ -1,0 +1,2 @@
+from .blog import *
+from .comment import *
