+++ kuhub/views/__init__.py
@@ -1,0 +1,2 @@
+from .blog_view import *
+from .comment_view import *
